+++ AutoRouteManager.php
@@ -14,6 +14,9 @@ namespace Symfony\Cmf\Component\RoutingAuto;
 
 use Symfony\Cmf\Component\RoutingAuto\AdapterInterface;
 use Symfony\Cmf\Component\RoutingAuto\Model\AutoRouteInterface;
+use Symfony\Cmf\Component\RoutingAuto\RoutingAutoEvents;
+use Symfony\Cmf\Component\RoutingAuto\Event\UriContextEvent;
+use Symfony\Component\EventDispatcher\EventDispatcher;
 
 /**
  * This class is concerned with the automatic creation of route objects.
@@ -22,30 +25,54 @@ use Symfony\Cmf\Component\RoutingAuto\Model\AutoRouteInterface;
  */
 class AutoRouteManager
 {
+    /**
+     * @var AdapterInterface
+     */
     protected $adapter;
+
+    /**
+     * @var UriGeneratorInterface
+     */
     protected $uriGenerator;
+
+    /**
+     * @var DefunctRouteHandlerInterface
+     */
     protected $defunctRouteHandler;
 
+    /**
+     * @var EventDispatcher
+     */
+    protected $eventDispatcher;
+
+    /**
+     * @var UriContextCollection[]
+     */
     private $pendingUriContextCollections = array();
 
     /**
      * @param AdapterInterface             $adapter             Database adapter
      * @param UriGeneratorInterface        $uriGenerator        Routing auto URL generator
      * @param DefunctRouteHandlerInterface $defunctRouteHandler Handler for defunct routes
+     * @param EventDispatcher              $eventDispatcher     Dispatcher for events
      */
     public function __construct(
         AdapterInterface $adapter,
         UriGeneratorInterface $uriGenerator,
-        DefunctRouteHandlerInterface $defunctRouteHandler
+        DefunctRouteHandlerInterface $defunctRouteHandler,
+        EventDispatcher $eventDispatcher
     )
     {
         $this->adapter = $adapter;
         $this->uriGenerator = $uriGenerator;
         $this->defunctRouteHandler = $defunctRouteHandler;
+        $this->eventDispatcher = $eventDispatcher;
     }
 
     /**
-     * @param object $document
+     * Build the URI context classes into the given UriContextCollection
+     *
+     * @param UriContextCollection $uriContextCollection
      */
     public function buildUriContextCollection(UriContextCollection $uriContextCollection)
     {
@@ -57,16 +84,7 @@ class AutoRouteManager
             $autoRoute = null;
 
             if ($existingRoute) {
-                $isSameContent = $this->adapter->compareAutoRouteContent($existingRoute, $uriContext->getSubjectObject());
-
-                if ($isSameContent) {
-                    $autoRoute = $existingRoute;
-                    $autoRoute->setType(AutoRouteInterface::TYPE_PRIMARY);
-                } else {
-                    $uri = $uriContext->getUri();
-                    $uri = $this->uriGenerator->resolveConflict($uriContext);
-                    $uriContext->setUri($uri);
-                }
+                $this->handleExistingRoute($existingRoute, $uriContext);
             }
 
             if (!$autoRoute) {
@@ -75,15 +93,45 @@ class AutoRouteManager
             }
 
             $uriContext->setAutoRoute($autoRoute);
+
+            if ($this->eventDispatcher) {
+                $event = new UriContextEvent($uriContext);
+                $this->eventDispatcher->dispatch(RoutingAutoEvents::URI_CONTEXT_BUILT, $event);
+            }
         }
 
         $this->pendingUriContextCollections[] = $uriContextCollection;
     }
 
+    /**
+     * Process defunct (no longer used) routes
+     */
     public function handleDefunctRoutes()
     {
         while ($uriContextCollection = array_pop($this->pendingUriContextCollections)) {
             $this->defunctRouteHandler->handleDefunctRoutes($uriContextCollection);
+        }
+    }
+
+    /**
+     * Handle the case where the generated path already exists.
+     * Either if it does not reference the same content then we 
+     * have a conflict which needs to be resolved.
+     *
+     * @param Route $route
+     * @param UriContext $uriContext
+     */
+    private function handleExistingRoute($existingRoute, $uriContext)
+    {
+        $isSameContent = $this->adapter->compareAutoRouteContent($existingRoute, $uriContext->getSubjectObject());
+
+        if ($isSameContent) {
+            $autoRoute = $existingRoute;
+            $autoRoute->setType(AutoRouteInterface::TYPE_PRIMARY);
+        } else {
+            $uri = $uriContext->getUri();
+            $uri = $this->uriGenerator->resolveConflict($uriContext);
+            $uriContext->setUri($uri);
         }
     }