+++ RoutingAutoEvents.php
@@ -1,0 +1,20 @@
+<?php
+
+/*
+ * This file is part of the Symfony CMF package.
+ *
+ * (c) 2011-2014 Symfony CMF
+ *
+ * For the full copyright and license information, please view the LICENSE
+ * file that was distributed with this source code.
+ */
+
+namespace Symfony\Cmf\Component\RoutingAuto;
+
+class RoutingAutoEvents
+{
+    /**
+     * Dispatched when the UriContext has been built
+     */
+    const URI_CONTEXT_BUILT = 'cmf_routing_auto.uri_context_built';
+}