+++ CHANGELOG.md
@@ -13,6 +13,7 @@ dev-master
                   instead of or in addition to URI strings.
 * **2015-01-29**: Added SymfonyContainerParameterTokenProvider for retrieving
                   path elements from the DI container
+* **2014-12-07**: Added event dispatcher and URI_CONTEXT_BUILT event
 
 1.0
 ---
@@ -20,3 +21,4 @@ dev-master
 * **2014-08-19**: Added validation for yaml routing auto configuration keys
 * **2014-08-10**: Replaced all instances of the term URL with URI
 * Separated this package from the RoutingAutoBundle
+